[iOS/macOS] Wide event local validation (#3663)
<!--
Note: This template is a reminder of our Engineering Expectations and
Definition of Done. Remove sections that don't apply to your changes.

⚠️ If you're an external contributor, please file an issue before
working on a PR. Discussing your changes beforehand will help ensure
they align with our roadmap and that your time is well spent.
-->

Task/Issue URL:
https://app.asana.com/1/137249556945/task/1213372223695149?focus=true
Tech Design URL:
CC:

### Description

This PR wires up the initial local validation stubs for iOS and macOS.
Right now, no actual validation occurs, but we now have the log file
available needed to make it happen. This **only** happens in `DEBUG`
mode.

### Testing Steps
<!-- Assume the reviewer is unfamiliar with this part of the app -->
1. Run the iOS app and send a wide event (you can do this by simply
signing into Privacy Pro), then run `./scripts/validate_wide_events.sh`
while the sim is still active - this will emit the logged wide events
2. Do the same for macOS

<!-- 
### Testability Challenges
If you encountered issues writing tests due to any class in the
codebase, please report it in the [Testability Challenges
Document](https://app.asana.com/1/137249556945/project/1204186595873227/task/1211703869786699)

1. **Check the Document:** First, check the **Testability Challenges
Table** to see if the class you encountered is listed.
2. **If the Class Exists:**
   - Update the **Encounter Count** by increasing it by 1.
3. **If the Class Does Not Exist:**
   - Add a new entry
-->

### Impact and Risks
<!-- 
What's the impact on users if something goes wrong?

High: Could affect user privacy, lose user data, break core
functionality
Medium: Could disrupt specific features or user flows
Low: Minor visual changes, small bug fixes, improvement to existing
features
None: Internal tooling, documentation
-->

#### What could go wrong?
<!-- Describe specific scenarios and how you've addressed them -->

### Quality Considerations
<!-- 
Focus on what matters for your changes:
- What edge cases exist?
- How does this affect performance?
- What monitoring have you added?
- What documentation needs updating?
- How does this impact privacy/security?
-->

### Notes to Reviewer
<!-- Anything specific you want reviewers to focus on -->

---
###### Internal references:
[Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f) |
[Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552) |
[Tech Design
Template](https://app.asana.com/0/59792373528535/184709971311943)

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> **Medium Risk**
> Mostly debug-only logging and local tooling, but it also changes
wide-event duration/bucketing types and updates wide-event JSON schemas,
which could affect metrics shape/values if mismatched with downstream
expectations.
> 
> **Overview**
> Adds **DEBUG-only local validation logging** for pixels and wide
events: `PixelKit` now writes fired pixel URIs to
`pixelkit-validation-log.txt`, and `DefaultWideEventSender` writes JSONL
wide-event payloads to `wide-event-validation-log.jsonl`.
> 
> Standardizes wide-event latency handling by switching
`MeasuredInterval.durationMilliseconds`/`DurationBucket` from `Double`
to `Int`, updating subscription wide-event bucketing helpers
accordingly, and tweaks wide-event pixel logging messages.
> 
> Updates wide-event JSON schemas to constrain `sample_rate` to `[0,1]`
and include optional `is_first_daily_occurrence`, bumps
`@duckduckgo/pixel-schema` to `v2.1.0`, and adds/extends iOS/macOS
scripts and npm commands to validate captured pixel and wide-event logs.
> 
> <sup>Written by [Cursor
Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit
4a96eb442307dfc76ea1bfc13268ebb5c85d4421. This will update automatically
on new commits. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

---------

+++ SharedPackages/BrowserServicesKit/Sources/PixelKit/WideEvent/WideEventData.swift
@@ -338,9 +338,9 @@ extension WideEventErrorData: WideEventParameterProviding {
 
 extension WideEvent.MeasuredInterval {
 
-    public var durationMilliseconds: Double? {
+    public var durationMilliseconds: Int? {
         guard let start, let end else { return nil }
-        return max(end.timeIntervalSince(start) * 1000, 0)
+        return max(Int(end.timeIntervalSince(start) * 1000), 0)
     }
 
     public func stringValue(_ bucket: DurationBucket) -> String? {
@@ -355,12 +355,12 @@ extension WideEvent.MeasuredInterval {
 
 public enum DurationBucket {
     case noBucketing
-    case bucketed((Double) -> Int)
+    case bucketed((Int) -> Int)
 
-    func apply(_ ms: Double) -> Int {
+    func apply(_ ms: Int) -> Int {
         switch self {
         case .noBucketing:
-            return Int(ms)
+            return ms
         case .bucketed(let fn):
             return fn(ms)
         }
