+++ SharedPackages/BrowserServicesKit/Sources/Subscription/WideEvents/SubscriptionPlanChangeWideEventData.swift
@@ -154,7 +154,7 @@ extension SubscriptionPlanChangeWideEventData {
         self.errorData = WideEventErrorData(error: error)
     }
 
-    private static func bucket(_ ms: Double) -> Int {
+    private static func bucket(_ ms: Int) -> Int {
         switch ms {
         case 0..<1000: return 1000
         case 1000..<5000: return 5000
