+++ SharedPackages/BrowserServicesKit/Sources/Subscription/WideEvents/SubscriptionPurchaseWideEventData.swift
@@ -148,7 +148,7 @@ extension SubscriptionPurchaseWideEventData {
         self.errorData = WideEventErrorData(error: error)
     }
 
-    private static func bucket(_ ms: Double) -> Int {
+    private static func bucket(_ ms: Int) -> Int {
         switch ms {
         case 0..<1000: return 1000
         case 1000..<5000: return 5000
